http-server: dont crash on WS close message
